app/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
whitenoise==5.3.0
